Déplace la vue member_from_ip
